Update Jenkinsfile - stop 3

+++ Jenkinsfile
@@ -51,16 +51,20 @@ pipeline {
         }
         stage('Stop and Remove Container') {
             steps {
-                // Get the container ID
-                CONTAINER_ID = sh(script: "docker ps -aqf 'ancestor=stock_dashboard'", returnStdout: true).trim()
+                // Get the running container ID
+                CONTAINER_ID = sh(script: "docker ps -qf 'ancestor=stock_dashboard'", returnStdout: true).trim()
 
-                // Stop the Docker container
-                sh "docker stop ${CONTAINER_ID}"
-
-                // Remove the Docker container
-                sh "docker rm ${CONTAINER_ID}"
+                // Check if the container ID is not empty
+                if (CONTAINER_ID) {
+                    // Stop the Docker container
+                    sh "docker stop ${CONTAINER_ID}"
+                    // Remove the Docker container
+                    sh "docker rm ${CONTAINER_ID}"
+                } else {
+                    echo "No running container found."
+                }
             }
-        }      
+        }     
     }    
 }
